Test: testing

+++ app/controllers/ApiController.php
@@ -28,10 +28,15 @@ class ApiController
 
         
         if (hash_equals('sha1=' . $hash, $githubSignature)) { 
-            file_put_contents('./logs/auto/test.log', 'githubSignature:' . $githubSignature . ";\n" . 'hash: sha1=' . $hash . "\n\n\n coucou \n\n\n");
             // On integre un message de validation avec la date et l'heure
             $date = date('d/m/Y H:i:s');
             $data = $date . ' - ' . $payload;
+
+            // On regarde si le script est présent dans le dossier "app/auto"
+            if(!file_exists('./app/auto/autodeploy.sh')) {
+                file_put_contents('./logs/auto/test.log', "existe pas\n", FILE_APPEND);
+                return;
+            }
 
             // On enregistre le payload dans un fichier
             file_put_contents('./logs/auto/payload.log', 'Valid payload:' . $data . ";\n", FILE_APPEND);
